+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Add support to dynamically set the number of decimal digit of BOM quantities
 - If there is no template defined for the object, generate an empty draft message for MESSAGE : wizard
 - Put buttons in hidden mode instead of readOnly in invoicing-project-form and put status to invoiced on generating invoice
+- Manage waste rate in bill of material and take it into account on cost sheet
 
 ## Improvements
 - Label "hours" on Project and Task with the field totalPlannedHrs.